admin: expose common admin server helpers
